fix(sdds-acore): Shapeable interface were moved to public area. Gradient generator were fixed.
